backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
# optional, only if /api/transcribe is used:
# openai-whisper==20231117
